Add multi-version rustdoc support: v16 (stable) and v21 (beta).
