+++ Cargo.toml
@@ -14,7 +14,7 @@ categories = ["command-line-utilities", "development-tools::cargo-plugins"]
 
 [dependencies]
 trustfall_core = "0.0.4"
-rustdoc-types = "0.17.0"
+trustfall_rustdoc = { version = "0.2.1", features = ["v16", "v21"] }
 clap = { version = "3.2.8", features = ["derive", "cargo"] }
 serde_json = "1.0.82"
 anyhow = "1.0.58"
